clean views and apis folders
