+++ srcs/django_code/transcend/chat/api/Groups.py
@@ -1,0 +1,25 @@
+from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.parsers import JSONParser, MultiPartParser, FormParser
+from rest_framework.response import Response
+from rest_framework import status
+from chat.models import *
+from chat.serializer import *
+
+class GroupsView(APIView):
+    permission_classes = (IsAuthenticated,)
+    parser_classes = [JSONParser, MultiPartParser, FormParser]
+    serializer_class = GroupSerializer
+
+    def get(self, request):
+        groups = Group.objects.filter(members=request.user)
+        serializer = GroupSerializer(groups, many=True)
+
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    def post(self, request):
+        serializer = GroupSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save(owner=request.user, members=[request.user])
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
