+++ srcs/django_code/transcend/game/views/pong.py
@@ -1,0 +1,5 @@
+from django.http import HttpRequest
+from transcend.views.utils import spa_render
+
+def pong(request: HttpRequest):
+    return spa_render(request, 'pong.html')
