+++ srcs/django_code/transcend/game/views/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .pong import *
+
+urlpatterns = [
+	path('pong/', pong, name='pong')
+]
