+++ srcs/django_code/transcend/transcend/views/home.py
@@ -1,0 +1,5 @@
+from django.http import HttpRequest
+from transcend.views.utils import spa_render
+
+def home(request: HttpRequest):
+    return spa_render(request, 'home.html')
