+++ srcs/django_code/transcend/transcend/views/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from .home import *
+from .sign import *
+
+urlpatterns = [
+	path('', home, name='home'),
+	path('sign', sign, name='sign')
+]
