+++ srcs/django_code/transcend/users/views/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .profile import *
+
+urlpatterns = [
+	path('user/<str:username>', profile, name='user')
+]
